# specs/001-student-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Student Management System API
  version: 1.0.0
  description: REST API for student enrollment, viewing, and record management
  contact:
    name: Student Management System
    email: admin@schoolsystem.com
servers:
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/v1/grades:
    post:
      summary: Record a grade for a student in a course
      operationId: recordGrade
      tags:
        - grades
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeCreate'
      responses:
        '201':
          description: Grade successfully recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
        '400':
          description: Bad request - validation failed (e.g., grade outside 0-100 range)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - teacher not assigned to course
  /api/v1/grades/{grade_id}:
    put:
      summary: Edit an existing grade (grade value, type, or dates)
      operationId: updateGrade
      tags:
        - grades
      security:
        - bearerAuth: []
      parameters:
        - name: grade_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeUpdate'
      responses:
        '200':
          description: Grade successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
        '400':
          description: Bad request - date order violation or student no longer on the course
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - the grade belongs to a course the teacher does not teach
        '404':
          description: Grade not found
  /api/v1/students/{student_id}/grades:
    get:
      summary: Get grades for a student
      operationId: getStudentGrades
      tags:
        - grades
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: courseId
          in: query
          required: false
          schema:
            type: string
            format: uuid
            description: Filter grades by course
      responses:
        '200':
          description: Successful response with student grades
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GradeDetail'
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/courses/{course_id}/grades:
    get:
      summary: Get all grades for a course
      operationId: getCourseGrades
      tags:
        - grades
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with course grades
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GradeDetail'
        '404':
          description: Course not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/v1/reports/{student_id}:
    get:
      summary: Generate a report card / progress report PDF for a student
      operationId: generateReportCard
      tags:
        - reports
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: >
            Successful response with a PDF. Admins get the full "Academic
            Report Card" for any student (filename `report_card_…`). Teachers
            get an "Academic Progress Report" scoped to their own courses only
            (filename `progress_report_…`); a teacher not teaching the student
            gets 403.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - teacher does not teach this student via any course
  /api/v1/reports/portal/{student_id}:
    get:
      summary: Generate report card for a parent's own child (PDF export)
      operationId: generateReportCardPortal
      tags:
        - reports
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with PDF report card
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (parent must be linked to this student)
...
  /api/v1/settings/brand:
    get:
      summary: Public school branding (name, tagline, colors, demo flag, logo URL)
      operationId: getBrand
      tags:
        - settings
      responses:
        '200':
          description: Brand settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandSettings'
        '500':
          description: Branding misconfigured (e.g. configured logo file missing)
  /api/v1/settings/brand/logo:
    get:
      summary: Public school logo image (also used as the browser favicon)
      operationId: getBrandLogo
      tags:
        - settings
      responses:
        '200':
          description: Logo image (image/png, image/jpeg, or image/webp)
          content:
            image/png:
              schema:
                type: string
                format: binary
        '500':
          description: Branding misconfigured (e.g. configured logo file missing)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GradeCreate:
      type: object
      required:
        - student_id
        - course_id
        - grade_value
        - date_assigned
        - date_due
      properties:
        student_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        grade_value:
          type: number
          minimum: 0
          maximum: 100
        assignment_type:
          type: string
          enum: [quiz, test, homework, final]
        date_assigned:
          type: string
          format: date
        date_due:
          type: string
          format: date
    GradeDetail:
      type: object
      properties:
        grade_id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        grade_value:
          type: number
        assignment_type:
          type: string
        date_assigned:
          type: string
          format: date
        date_due:
          type: string
          format: date
        date_graded:
          type: string
          format: date
    GradeUpdate:
      type: object
      description: Partial update of a recorded grade. Student and course are not editable.
      properties:
        grade_value:
          type: number
          minimum: 0
          maximum: 100
        assignment_type:
          type: string
          enum: [quiz, test, homework, final]
        date_assigned:
          type: string
          format: date
        date_due:
          type: string
          format: date
    GradeResponse:
      type: object
      properties:
        success:
          type: boolean
        grade:
          $ref: '#/components/schemas/GradeDetail'
        message:
          type: string
    BrandSettings:
      type: object
      properties:
        school_name:
          type: string
        tagline:
          type: string
        primary_color:
          type: string
          example: '#0B6B4F'
        secondary_color:
          type: string
          example: '#D9A441'
        demo:
          type: boolean
        logo_url:
          type: string